Rosters, Multi Selection, Command Profiles
Further expanded selection logic and accompanying UI to allow selection of multiples of different types of units.

When selecting units a card will appear with the name & count. The first Card will be the primary, and all commands will be loaded from that. Press Tab to switch between cards, UI will load the commands for that type. Command switching works, and will apply to the command to all units, regardless of which type is primary. If a unit cannot use a command, they will safely discard it.

Added basic Prefab Registry for Units and updated Selection event to match new selection logic.