Event Agent, Local & Global Events & Fixes
Minor fixes to Player controls, selection pane plays nicer with less errors. Created Event Agents to separate local and global events on the Event Bus. Subscribe to either an Event type for global events or an Agent ID for local events from that Agent. Call Local on the EventBus to post a local event and Global to post a global one.