Rudimentary Turret
Created a rudimentary turret for the Tall Tank and created a custom Tank class to be able to set targets. Use Attack with the Tall Tank and select another unit for the tank to attack. The tank will move towards the target and fire as soon as the target is in range. Tank will not attack at will.