ITaggable, Entities, Init Events
Changed the UnitCache to an EntityCache which registers a separate Entity Component. In order for objects to be referenced in the cache they now need to implement the ITaggable interface which in the future will require methods for handling Tags as a serialized data transfer method between components and objects, similar in style to Minecraft Forge's Capabilities system.

Refactored the current Unit classes to use this interface and work with the new cache, added Init events and implemented the EntityInit event for the use of turrets so that they only activate their colliders when entity registration is complete and they can reference entities via the cache. No actual change to gameplay so far.