Factory, Registry Expansion, Command Expansion
Created Command Pages to expand commands and have more explicit control over the placement of commands in the UI, created the construction commands though haven't fully implemented them. Created the first building, a Factory, as well as the ghost for it. Slowly building the logic for placing buildings using the commands.

Expanded the Registry to be an Abstract Prefab Registry that's got an implementation for Units & Buildings now.

Created the Constructor unit which can repair units and has the construct command, however neither are fully implemented yet.