Building Placement, Damage & Health
Implemented building placement by using the construct command to open up a command page including the factory building in it. Construction menus must be set up on the command object rather than on the unit. Building can be placed however the repair command hasn't been fully implemented. Building starts at a scale of 0, haven't implemented the scale increasing as it gets built yet.

Implemented simple health with the ISelectable interface & an attack function on the interface, repair turret currently just uses the Attack function with a negative damage value to heal the unit. Units attack, projectiles have a callback function in which a simple lambda is being used for now with hopes to implement onHit events.

Moved unit models to the Entity Layer to better control collisions.