Building Selection
Expanded the Roster to be able to accept any kind of registry unit, started on an IRegistryObject but wasn't able to get it working, currently implemented on the Unit and Building but not actually used.

Can now select the Factory after building it without any issues. A lot of work for a seemingly small change but in theory you should be able to accept any number of different units using ISelectable.