Production Command & Bug Fixes
Fixing some bugs with Factory logic, especially how it updates its commands after completing construction. Fixed an issue with Factory destruction, since the repair command was being delivered as soon as the factory was placed the unit was unable to subscribe to the death event. Fixed it by making it deliver the command on Entity Init.

Created the Produce command with its own implementation of Commandlet, still working on the factory side logic for production.