Unit Production & UI Improvements
Finished the first iteration of Unit Production, factories can be constructed by constructors. The factory can produce a constructor unit.

Lots of little UI improvements were made to make this possible, a lot of alterations were made to Health Bars, improvements to the events that drive them, to allow much better functionality and accuracy of information.

Also added hovering, edited the box selection to highlight units within its square as well as to highlight them when the mouse is hovering over them. Health bars will show according to if a unit is damaged or is hovered, the selection rings indicate that you're hovering a unit.

Added a production bar & improved the construction bar using some cheeky lambdas as subscribers to event bus events pushed by the factory.