Harvester Implementation
Created the first implementation for the Harvester Unit, Turret & Harvest Command as well as the ResourceDeposit implementation.

Refactored the turret system by removing the Raycast turret and replacing it with the AbstractTurret class, now used by the ProjectileTurret, BuilderTurret & HarvesterTurret implementations.

It better organizes the data the turret uses by using casts between ISelectable and the other sister interfaces. Currently the Harvester doesn't move towards the deposit to harvest it.