Depositing & UI
Created logic for the Harvester to return to the nearest IDepositable and deposit resources into the Player's banks. When the Harvester is full it'll return to the Scrapyard, remembering the deposit it was harvesting from, and deposit the resource units before returning to the deposit to harvest some more.

Created the Resource Panel UI element for each registered resource to the player, allowing for multiple resources to be created, which'll update whenever a resource is deposited.

Added costs to production commands which are refunded when the command is cancelled.

Lots of little fixes along the way. More Events created.