Rendered Icons
Created an Icon Renderer scene that uses a render texture to create icon sprites for each of the units. Currently a manual process as you have to manually tweak the viewport size & unit positions but they look nice.

Implemented them using the ISelectable to get an Icon, UI will now create unit cards using this icon by referring to the Registry. Modified the registry a bit more to be able to return implied types instead of just strict types.