Complete Unit Info
Continued expanding out the Unit Info as a part of the Unit Pane. Created implementations for:
Selection Name
Unit Health
Resource Deposit Value
Production Queue

each of these are activated by calling the Unit Info card and calling Get on the needed module, using a string as a key for each. Get the specific component you need and enter your values, all modules have event listeners that update the module depending on the current selection; won't activate when not in use.

CSS is NOT my passion...