Command Icons & Fixes
Lots of bug fixing, fixed Harvester forgetting their harvest target and being unable to harvest, fixed the bug with buildings exploding in size, fixed Unit command flow so that targets are properly reset when commands are cancelled, fixed the UI failing to update on building death, fixed the cancel construction editing the Player's selected dictionary before the Select function was done with it.

Created simple icons for all commands that're displayed in the panel (so deposit doesn't have one for example), production & construction icons use the icon of their production/construction.

Implemented cost for building constructions, cancelling construction will refund the cost back to the player.