Oil, Building Placement Checks
Started implementing Oil as a resource. Began with the Oil Slick deposit and Pumpjack building. The pumpjack is needed to extract the oil, it itself is a harvestable that needs to be harvested by a (yet to be implemented) tanker unit.

Altered the building placement script to add legality checks, most buildings will check this based on if there's something in the way. The pumpjack's placement script is modified to snap the pumpjack to the oil deposit closest to the cursor, and the will determine legality based on if it's over one or not. Materials update to red or green to indicate legality.