Command Tooltip
Created the Oil Resource and made the Tanker deposit Oil into the Oil resource instead of just Resource Units.

Created the Command Tooltip to display the cost & description of all commands, currently a very simple implementation that doesn't actually work at the moment. Currently the tooltip displays all the information in a neat and ordered manner and attempts to resize the tooltip according to the amount of information provided however this currently isn't behaving as I expect. Gave Commands descriptions to do this.