Finished Tooltip
Tooltip behaves correctly now, found that adding true to both the bool parameters in ForceUpdateMesh for text meshes fixed both the sizing issues we had.

Added a Cancel Command function to commands for when you select between commands in the UI and prevented ambiguity between clicking on UI & the world by moving left click input to a separate function which checks if we're hovering the UI, and if so to then cancel the call.