Fog Stencil
Started on Fog of War by experimenting with Render Textures and shaders to create a fog stencil. Was sucessfully able to create a static fog that is less dark after exploring, and a dynamic fog that completely returns when not visible.

Started work on a more data oriented solution in order to better integrate it with systems and have better control over the stenciling, most of the work today was getting the rendering side working.