Entity Vision
Created algorithms for an Entity to calculate its vision while accounting for blocked vision.

The method is quite complex and takes a lot of computation, I've made some optimizations that've made the performance cost on one unit unnoticable, but will work on further optimizing based on testing.

Algorithm works by collecting all vision nodes that're in range of the vision and separates them according to their distances. It will then sequentially go through each outer layer, working inwards. It will perform a raycast from the center to the currently inspected node, and then set the visible nodes according to how far that raycast travels.

When it does this, it removes all nodes marked as visible/blocked from the open set, reducing the amount of nodes inspected significantly.

This updates per unit every second using a Coroutine to keep it running. I reduced the resolution of the grid to one node per 2 Unity units to further optimize it.

Ideas for further optimization:
- Move processing to Vision Singleton & Multi-Thread it
- Refine the algorithm to further reduce the amount of nodes inspected
- Use Jobs system