Vision Renderer
Created a Vision Renderer component as a second component for the main Vision singleton. This renders a texture on the CPU and sends it to the GPU to be rendered with the help of the higher resolution render texture.

Original version was very unoptimized, calculating for all nodes every frame losing about 180 fps. Created a Dirty check which fixed the frame loss but caused a noticable hitch. Identified that the biggest performance cost was calculating the color and creating the color array, so I moved that to a separate thread as well which has fixed the performance issues with the system.

Want to explore the possibility of a Compute Shader to calculate this.

Also fixed occasional vision loss bug.