Unit Hiding
Implemented the Vision system with the rest of the game. Units & Buildings will now update their visibility according to the visibility in the vision system. This involves currently just looping through a serialized array of objects that need to be hidden/shown when vision changes. This disables all colliders, added a ground collider so units would stop falling through the floor.

All units will be completely hidden when not in vision range, however buildings will remain completely visible and selectable if it's been discovered before.

Added multiple events to implement, listen for the global VisionInit to know when the vision system is ready to be used, and global Vision Updates to know when a unit/building should check if it's visible to update.

Turned the SelectionCircle into its own script & a prefab to accompany to support this more intuitively.