Infantry Squad Implementation
Created the Infantry squad and implemented the first version of Infantry. Infantry work with a clever use of selectable colliders, each selectable collider is a child object of the parent squad which doesn't move. Every frame it matches each collider with each infantry unit, allowing for selection of just the squad. The squad will then distribute commands to each individual in the squad who will act independently as units.

The units still maintain the rest of the colliders to function for interactions between units, allowing them to have independent vision and still be attacked.

Currently, infantry only move with attacking to be implemented soon to follow.

Some changes were made to the selection pane to make this possible so that multiple selectable colliders that reference the same unit don't result in multiple selection entries, instead resulting in only 1.

Adjusted the selection circle sprites so that infantry's circles are more visible.