Roughneck's & Stealth
Altered the IsVisible function to first query a registered vision component if one can be found. Vision components now have a VisibleTo property which contains the bitmask of all player's that unit is visible to.

This was done so that stealth for the Roughnecks unit was a possibility. Roughnecks can now toggle a sneak command which will make them hidden to any units unless they get within a certain range, indicated to the player with a range sprite.

Implemented a couple events to help do this and created a derived class of the EntityVision class specifically for the Roughnecks.