Infantry Attacking & Building
Implemented attacking for the Roughneck squad by making a special version of the projectile turret for them which fires in bursts, doing only a small amount of damage. This wont fire when within stealth.

Created the Makeshift Pumpjack which can be built by the Roughnecks, commands are in place for Roughnecks to harvest and deposit but still implementing.

Lots of small fixes all over the place.