Mobile Artillery
Created the Mobile Artillery which needs to deploy before firing. It's turret will listen out for a DeployEvent called on the unit before revealing it's range indicator and firing on targets in view. This now comes with a much smaller gatling turret which will constantly fire whether deployed or not, but does a very small amount of damage.