Command Reworking
Moved command queues to a component so that we don't have to keep custom creating command queues on each unit's class, massively inflating their classes. Mainly working to simplify and standardize commands a bit so that we can understand data about them. The goals of this rework are to allow the command panel UI to understand:
 - Command Cooldowns
 - Command Usability
 - Command/Production Progress
 - Command Toggle State

Since we have all of these currently in the game I want the UI to be able to deliver this feedback to the player, we're doing this by moving everything to a component that reveals commands with active cooldowns or an active toggle state, allowing the UI to understand how to display each command button. So far move command is the only one that works, infantry needs a rework on how it handles commands and cannot do anything.