Debugging, Minor Refactor
Still debugging the command queue changes, infantry member has a really weird bug that doesn't actually break behaviors but throws a lot of errors; the first member that reaches the end of a move command will call OnPathComplete, completing the command, but for some reason posting the event doesn't return the stack to the same point in the function, re-running it on null objects that we no longer have references to, exiting that member's script and preventing them from setting the path to empty. All other members behave accordingly, it's just this first one that seems to trigger the call twice for some reason, stack trace hasn't helped in locating why.