Fixed Infantry Bug & UI Work
Fixed the bug with infantry spitting out errors when cancelling move orders for another. Played around with the UI to create effects for active commands, disabled buttons and a cooldown overlay that unfills the button in a clockwise fashion.

Added a Cooldowns list to the Command queue and implemented the Active list to the ICommandable interface. Got most of the behavior of the Sneak Command syncing up units to use the ability.