Production Conversion
Implemented the Production Queue and implemented it for Buildings, using it for both upgrades and regular productions. Created the IProducable interface to assist with this, any upgrade cannot be re-commanded after being queued up. Production mostly works, there's some quirkiness with a delay before it gets queued up properly, causing minor issues with the UI information.