Bug Fixing
Fixed some bugs related to the queue and how info is displaying. Fixed the delay as well as failing to update the UI for the whole production queue. Fixed the upgrade production as well as another bug with stealth when approaching resource deposits. Adjusted selection circle masks but there's another bug where the mask is still active even if the circle is not.