Visibility Fix
Fixed buildings not consistently being revealed when they're supposed to be, as well as sensor targeting that I didn't know was broken until fixing the above. Finishing this commit now as I want to install URP and test out some blur using it, however it could potentially break the current shaders so would like to be able to revert it easily.

Currently, this branch is feature complete, once blur has been tested it'll be ready to be pulled into dev & stable.