URP And more Polish
Tried to use URP to get a cheap blur effect on the vision render, however discovered the method won't work due to using an orthographic render camera. Did however update everything to URP, updated all materials and recreated projector shader with a decal shader which worked great.

Little fixes, finished the DIstributeOrder in the Player to distribute production orders which works. Implemented ICommandable.Count to count the current command when checking how many commands are in a commandable's queue.