Camera Zoom & Height
Moved camera movement to its separate component in the ViewPortController and added an extra gameObject; the Gimbal.

I want look rotating to be orbital, so I moved the actual main player to 500 units above the map and it fires a ray down, moving the gimbal child object to the highest point on the walkable terrain it hits, keeping the camera at a consistent height.

Camera now has zooming, originally had this based on height but preferred doing it along Z axis. As the camera is a child object of the gimbal it will rotate around the gimbal, zooming in and out of where the gimbal is currently looking. Horizontal movement is handled by the root object.