Research
Started implementing research for the Adrenaline technology which needs to be unlocked at the Barracks for Marines to be able to use them. Technology is just done via gameobejcts with their names registered to a table that the main player tracks, researched technologies are then queried by the Marines at the central player object. Currently adrenaline isn't being unlocked by marines.