Bug Fixing
Fixed ability timer changes stacking, making timers more consistent in their use of cooldowns as well as for UI reactivity. Moved the check for adrenaline being made inactive to a check for if the cooldown is completed to achieve this as well as more reliable event posting when the cooldown is completed or not.

Sped up projectiles, made them collide with terrain and gave them a lifetime to significantly improve their reliability, speeds have been almost quadrupled so moved projectiles to continuous collision detection which has so far worked well.