Shaders, Batching & Viewport Rect
Created a simple line renderer to represent where the viewport rect is currently looking on the minimap, however there is currently a bug with how the raycast is calculating the points. After rotating the camera under 30 two of the rays for some reason fire backwards.

Played with shader graphs to try and get the selection circles to use batching, as well as the minimp icons which was mostly sucessful. Had to play with the URP settings to get it to work as the batching settings are strange.