Sniper & Minimap Interactions
Got interactions with the minimap working, left clicking on the minimap will move the camera to it, while maintaining the interpolation to keep it smooth. Right clicking with units selected will order them to move to the point on the minimap.

Started making the Sniper infantry unit, a sneaking, high damage long range and squishy unit. Meant to do bonus damage to other infantry, however currently does high damage to all units & buildings.