Flare, Sniper Implementation
Implemented the Sniper and created the Flare entity for it to spawn. The flare uses a vision component to give it's owner vision in a wide area for a short time before it destroys itself.

Had to modify how Death events are used by the Event Bus & Entity Cache by creating a separate Entity Destroy event which is called by the Entity component on object destroy. Created a simple icon and implemented the command for the flare but haven't just yet finished it.