Flare Implementation
Finished implementing the flare, giving it selectable & attackable interfaces (while not giving it colliders for units to interact with it) so that it can provide more information to the UI. Can select the flare which now shows a health bar that decays as the flares lifetime runs its course.

Finished implementing the flare command and the sniper unit using it, the sniper can fire the flare with a long cooldown giving it as much vision range as its weapon range which is two and a half times its vision range.