Bug Fixing & Landmine Prefab
Started on the landmines by making the construction prefab and starting on the logic for the actual mines. Need to work out logic between building 3 at once but each mine acting individually.

Fixed stealth circles not hiding when sneak is deactivated and the sniper not activating sneak in its command queue properly.