Landmine Implementation
Created an implementation for the landmines, which behave very similarly to infantry squads. The central landmine object contains its member mines as children in its prefab, which it then unparents, while tracking each mine individually. The mines then behave autonomously, using a trigger to detect if an enemy has stepped on it.

Selection colliders remain on the master object, showing all child mines health and combining their healths in the expanded unit info.

Implementation is still quite buggy, mines are being deleted as soon as another sensor detects its entity. Need to assign a separate rigidbody to the mine's detector.