Landmines Fixes
Fixed the landmines, completed their implementation which required some minor rejigging of building code. Did the icons for the mines & Sniper and made them both producable, Roughnecks can now build mines and snipers can now be produced from a Barracks.

Started work on a stealthsensor to subvert stealthed units. This is to allow the sniper to see stealthed units after acquiring an upgrade.