Stealth Sensor & Decoupling
Decoupled unit stealth from the entity vision component. Stealth is now a separate monobehavior. This was done so that other units can negate stealth units with their vision. The way this works is vision is now split into 3 stages:
 - Vision Update -> The GameVision calls this each update
 - Vision Check -> The vision component posts events to check the visibility
 - Entity Update -> The entity updates its rendering components

This way, the stealth intercepts the check in the Pre phase and overrides it with those in proximity, and the stealth sensor intercepts the check in the Post phase and adds its vision mask to the check.

Added a research to the barracks for this called thermalvision but haven't actually started making this yet.

Found a bug with attacking infantry squads where the incorporeal squad object is being targeted as the attackable rather than individual infantry members.