Thermal Vision Research
Implemented a thermal vision upgrade to the stealthdetectors that decides if they're detecting or not. Added a simple boolean property to the sensor to determine if its running which the upgrade simply sets on or off depending on if the tech has been researched.

Implemented RemoveListener to the global event bus.