Client Connection & New Scene
Starting with the basics of networking, working on just building each piece as I get to it. We got the client to successfully connect and spawn a player object using a faction which registers itself with a team on the server, not reflected on the clients yet. Testing this using Parallel Sync.

The host can press Start game to actually start the game, which will initialize the teams with one player per team and spawn a headquarters version of the scrapyard prefab. It will then trigger the vision to set the mask to the current player.

This so far works for the host, however not the clients as I haven't yet introduced networked behaviors to any of the objects. Created a simple gamestart button & GameInit object in the controller that initializes things, will refactor it properly when I know it's full scope.

Played around with assemblies to try and speed up compile time but that was a disaster so reverted it. Tried to sort namespaces into their own assemblies but as it turns out you can't have cyclical assembly dependencies so nothing worked.