Team Cache
Getting the Team Cache to initialize with game start and get players into their sorted teams. Not entirely successful, there's an issue with calling a Client RPC function which I suspect may be because it wasn't a NetworkBehavior the whole time :)))))

We learn, haven't had the chance to test it yet but will find out next time if that was the case. If not, it's likely a timing issue, requiring a handshake to verify each client's version of the game is adequately loaded before initializing the game.