Commandlet OBJs
Converting commandlets to use GameObjects and be network behaviors to allow syncing up of data and RPC calls to execute on each side. Currently got instantiation logic going with the RPCs however need to parse the selected unit data to the server so it knows what units to send the command to, then need to get the queue to RPC on the client whenever it enqueues/executes a command.