Command Serializers
So the wrappers needed to be expanded to a separate command serialization factory as you can't parse through subtypes of structs into a packet. Thus we had to serialize a key that points to a specifically registered serializer that can read and write command serializations. This now allows commands to properly sync up with custom data and do unique shit as the production command now mostly works.

The only thing to get to work is to properly use the OnWork event, then implement this for other commands. Once done, we can start on implementing netcode onto units.