Fixed RPC Serialization
Fixed the RPC serialization for the command as well as when the Player checks the owner of spawned units, now checking when the units ownership is initially set. However there's an issue with getting the NetworkObject from the EntityCache for some reason, preventing the selection from being passed through to the server.