Network Combat
Got unit combat working over the network by having turrets now be network behaviors. Updated all unit prefabs and the different turret types to appropriately work over the network, while adding more of the command prefabs. Still need to rework each individual command but we made a lotta progress today getting this to work, including networking player resources.

General convention that seems to work for network variables is to move events that're fired to responding the value change of the network variable so that the client is able to fire the event and make the necessary UI updates.