Harvest & Deposit
Discovered I can add more events to sensors & that I can use the attribute events to get rid of needing update in the receivers. We WANT an event driven paradigm, so moving more towards it from the update system actually really complements the patterns and smooths out the code. I'm VERY happy with the solution here.